Extracts classes from cdx_client.rb to lib directory
        The classes inline were getting unweildy, so I extracted
        the new classes into their own files in a lib directory.
