+++ cdx_client.rb
@@ -4,95 +4,8 @@ require 'forwardable'
 require 'savon-multipart'
 require_relative 'secret'
 
-module CDX
-  class Client
-    attr_reader :savon
-
-    def initialize(opts={})
-      @savon = Savon.client(default_opts.merge(opts))
-    end
-
-    def default_opts
-      {
-        :wsdl => "https://devngn.epacdxnode.net/cdx-register/services/RegisterAuthService?wsdl",
-        :pretty_print_xml => true,
-        :log => true,
-        :soap_version => 2,
-        :convert_request_keys_to => :none,
-      }
-    end
-
-    def self.signing
-      new({
-        :multipart => true,
-        :filters => [:password, :credential, :answer]
-      })
-    end
-
-    def self.auth
-      new({
-        :filters => [:password]
-      })
-    end
-
-    extend Forwardable
-
-    def_delegators :savon, :call
-
-    Signing = signing
-    Auth = auth
-  end
-
-  class User
-    attr_reader :output_stream, :opts
-
-    def initialize(opts, output_stream=$stdout)
-      @opts = opts
-      @output_stream = output_stream
-    end
-
-    def authenticate
-      puts opts
-      puts authentication_response.hash
-      repackage_response
-    end
-
-    private
-
-    def puts(message)
-      output_stream.puts(message)
-    end
-
-    def user_id
-      opts['userId']
-    end
-
-    def password
-      opts['password']
-    end
-
-    def user_data
-      authentication_response.hash[:envelope][:body][:authenticate_response][:user]
-    end
-
-    def authentication_response
-      @authentication_response ||= CDX::Client::Auth.call(:authenticate, {
-        :message => {
-          :userId => user_id, :password => password
-        }
-      })
-    end
-
-    def repackage_response
-      {
-        :UserId => user_data[:user_id],
-        :FirstName => user_data[:first_name],
-        :LastName => user_data[:last_name],
-        :MiddleInitial => user_data[:middle_initial]
-      }
-    end
-  end
-end
+require_relative 'lib/cdx/client'
+require_relative 'lib/cdx/user'
 
 def authenticate(args)
   signature_user = authenticate_user(args)
